requirements.txt
cryptography==41.0.5
Faker==19.13.0
firebase_admin==6.2.0
Flask==3.0.0
python-dotenv==1.0.0
Requests==2.31.0
rsa==4.9
